WaE: narrowing from int/long to float in initializer list (-Wc++11-narrowing)
Warnings from Clang now with -std=gnu++0x.

+++ vcl/aqua/source/a11y/aqua11ytextattributeswrapper.mm
@@ -98,7 +98,7 @@ using namespace ::rtl;
 
 +(void)addColor:(sal_Int32)salColor forAttribute:(NSString *)attribute andRange:(NSRange)range toString:(NSMutableAttributedString *)string {
     if ( salColor != -1 ) {
-        CGFloat elements[] = { salColor & 0x00ff0000, salColor & 0x0000ff00, salColor & 0x000000ff };
+        CGFloat elements[] = { static_cast<CGFloat>(salColor & 0x00ff0000), static_cast<CGFloat>(salColor & 0x0000ff00), static_cast<CGFloat>(salColor & 0x000000ff) };
         CGColorRef color = CGColorCreate ( CGColorSpaceCreateWithName ( kCGColorSpaceGenericRGB ), elements );
         [ string addAttribute: attribute value: (id) color range: range ];
         CGColorRelease ( color );
