+++ vcl/aqua/source/gdi/salgdiutils.cxx
@@ -189,7 +189,7 @@ bool AquaSalGraphics::CheckContext()
 
         if( !mrContext )
         {
-            const CGSize aLayerSize = {nWidth,nHeight};
+            const CGSize aLayerSize = { static_cast<CGFloat>(nWidth), static_cast<CGFloat>(nHeight) };
             NSGraphicsContext* pNSGContext = [NSGraphicsContext graphicsContextWithWindow: mpFrame->getWindow()];
             CGContextRef xCGContext = reinterpret_cast<CGContextRef>([pNSGContext graphicsPort]);
             mxLayer = CGLayerCreateWithContext( xCGContext, aLayerSize, NULL );
