+++ vcl/aqua/source/gdi/salprn.cxx
@@ -229,7 +229,7 @@ void AquaSalInfoPrinter::setPaperSize( long i_nWidth, long i_nHeight, Orientatio
     }
     else if( i_nWidth > 0 && i_nHeight > 0 )
     {
-        NSSize aPaperSize = { TenMuToPt(i_nWidth), TenMuToPt(i_nHeight) };
+        NSSize aPaperSize = { static_cast<CGFloat>(TenMuToPt(i_nWidth)), static_cast<CGFloat>(TenMuToPt(i_nHeight)) };
         [mpPrintInfo setPaperSize: aPaperSize];
     }
     // this seems counterintuitive
